Update codes according to reviews
